Move files to main directory
