+++ src/main/script.test.ts
